refactor tof driver hw test
