+++ packages/tof_driver/include/hardware_test_tof/__init__.py
@@ -1,0 +1,1 @@
+from .hardware_test_tof import HWTestToF
