+++ packages/tof_driver/setup.py
@@ -1,0 +1,11 @@
+## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+# fetch values from package.xml
+setup_args = generate_distutils_setup(
+    packages=["hardware_test_tof", "tof_accuracy"],
+    package_dir={"": "include"},
+)
+
+setup(**setup_args)
